+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "object-bystring",
-  "version": "2.1.3",
+  "version": "3.0.0",
   "description": "Access and alter objects using string literals",
   "main": "dist/index.js",
   "scripts": {
@@ -13,8 +13,11 @@
     "url": "git+https://github.com/tamb/object-bystring.git"
   },
   "keywords": [
-    "['object'",
-    "'bystring']"
+    "object",
+    "bystring",
+    "string",
+    "utility",
+    "mutable"
   ],
   "author": "tamb <tsaporito@fastmail.com>",
   "license": "MIT",
