+++ src/index.test.js
@@ -1,21 +1,25 @@
-require("./index.js");
+const byString = require("./index.js");
 const mockObject = require("./mocks").mockObject;
 
-test("Polyfill adds function to Object prototype", () => {
-  expect(Object.prototype.byString).toBeInstanceOf(Function);
+test("Object prototype remains unchanged", () => {
+  expect(Object.prototype.byString).not.toBeDefined();
+});
+
+test("byString to be a function", () => {
+  expect(byString).toBeInstanceOf(Function);
 });
 
 test("String gets variable from object", () => {
-  const firstName = mockObject.byString("name.first");
+  const firstName = byString(mockObject, "name.first");
   expect(firstName).toBe("John");
 });
 
 test("String sets variable in object", () => {
-  mockObject.byString("name.first", "James");
-  expect(mockObject.byString("name.first")).toBe("James");
+  byString(mockObject, "name.first", "James");
+  expect(mockObject.name.first).toBe("James");
 });
 
 test("String sets variable one deep", () => {
-  mockObject.byString("people[1].name", "Jeremy");
+  byString(mockObject, "people[1].name", "Jeremy");
   expect(mockObject.people).toBe("Jeremy");
 });
